[Core] findByExample NOK if IN array is empty #71
Multiple changes in this file.

1) 2 more tests added to "IN" case management
2) I noticed that if "IN" criteria array is empty, criteria is skipped
and in case this was the only criterie, we got a "WHERE " followed by
nothing and another DB error. So I added "WHERE 1" in select variable
and removed from the foreach.
3) With case 2, "AND " had the same issue : if no valid criteria, then
AND was the last part of the query -> db error. So I added "AND" at the
beginning of each criteria

I saw no other solution.